move SO multi-obs CLI to standalone so_mapmaking package
