+++ src/furax/interfaces/sotodlib/mapmaker.py
@@ -11,7 +11,7 @@ from sotodlib.core import AxisManager
 from sotodlib.preprocess.preprocess_util import init_logger
 from sotodlib.site_pipeline.utils.pipeline import main_launcher
 
-from furax.interfaces.sotodlib.observation import SOTODLibObservation
+from furax.interfaces.sotodlib import SOTODLibObservation
 from furax.mapmaking.config import MapMakingConfig
 from furax.mapmaking.mapmaker import MapMaker
