+++ src/so_mapmaking/run.py
@@ -1,13 +1,13 @@
 from pathlib import Path
 
-import jax
-
-from furax.interfaces.sotodlib.observation import LazySOTODLibObservation
-from furax.mapmaking import MapMakingConfig, MultiObservationMapMaker
+from cyclopts import App
 
 from .util import resolve_obsids, setup_logger
 
+app = App(help='Run the mapmaker on prepared binary observation files.')
 
+
+@app.default
 def run(  # type: ignore[no-untyped-def]
     obsdir: Path,
     obsid: list[str] | None = None,
@@ -28,6 +28,17 @@ def run(  # type: ignore[no-untyped-def]
         loglevel: Logging level (debug, info, warning, error).
         log_path: Log output path.
     """
+    # Defer JAX (and the mapmaker stack it pulls in) to call time so that
+    # merely importing this module stays cheap and backend-free.
+    import jax
+
+    from furax.distributed import maybe_init
+
+    maybe_init()  # must run before the JAX backend is touched
+
+    from furax.interfaces.sotodlib import LazySOTODLibObservation
+    from furax.mapmaking import MapMakingConfig, MultiObservationMapMaker
+
     logger = setup_logger(loglevel, log_path, process_index=jax.process_index())
 
     obsids = resolve_obsids(obsid, obsids_file)
@@ -70,3 +81,7 @@ def run(  # type: ignore[no-untyped-def]
     except Exception as e:
         logger.exception('mapmaking failed', exc_info=e)
         return 1
+
+
+if __name__ == '__main__':
+    app()
