+++ src/so_mapmaking/util.py
